requirements.txt
# paho-mqtt
# protobuf

paho-mqtt==1.6.1
protobuf==5.29.*
# git+https://github.com/eclipse/tahu.git@main  # Include the Tahu library
